requirements.txt
# ============================================
# Core Dependencies
# ============================================

# Deep Learning
torch==2.1.0
torchvision==0.16.0
onnx==1.15.0
onnxruntime-gpu==1.16.3

# Computer Vision
opencv-python==4.8.1.78
pillow==10.1.0
albumentations==1.4.0
face-recognition==1.3.0
retinaface==0.0.13

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Data Collection
yt-dlp==2023.11.16
instaloader==4.10.3
tiktok-downloader==1.0.0

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0

# Monitoring
tensorboard==2.15.1
wandb==0.16.1
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

